+++ src/main/java/com/develhope/spring/client/ClientService.java
@@ -2,21 +2,11 @@ package com.develhope.spring.client;
 
 import com.develhope.spring.client.clientControllerResponse.*;
 import com.develhope.spring.loginSignup.IdLogin;
-import com.develhope.spring.order.OrderRepository;
-import com.develhope.spring.rent.RentDto;
-import com.develhope.spring.order.dto.OrderClientDTO;
-import com.develhope.spring.order.dto.PurchaseClientDTO;
-import com.develhope.spring.order.OrderEntity;
-import com.develhope.spring.order.OrderType;
+import com.develhope.spring.order.*;
 import com.develhope.spring.rent.*;
-import com.develhope.spring.rent.RentStatus;
-import com.develhope.spring.seller.SellerEntity;
-import com.develhope.spring.vehicle.SellType;
-import com.develhope.spring.vehicle.VehicleEntity;
-import com.develhope.spring.rent.RentRepository;
-import com.develhope.spring.client.clientControllerResponse.*;
-import com.develhope.spring.seller.SellerRepository;
-import com.develhope.spring.vehicle.VehicleRepository;
+import com.develhope.spring.order.dto.*;
+import com.develhope.spring.seller.*;
+import com.develhope.spring.vehicle.*;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
